requirements.txt

Flask
fastapi[all]
uvicorn


pydub

PyAudio
numpy
scipy
matplotlib
mysql-connector-python
psycopg2-binary
python-multipart